(REF) API 서비스 분리 (#20)
