+++ app/src/main/java/com/gmg/seatnow/domain/repository/AuthRepository.kt
@@ -40,9 +40,6 @@ interface AuthRepository {
     suspend fun ownerLogout(): Result<Unit> // 로그아웃
     suspend fun ownerWithdraw(businessNumber: String, password: String): Result<Unit> // 회원탈퇴
     suspend fun withdrawUser(): Result<Unit>
-    suspend fun getOwnerAccount(): Result<OwnerAccountResponseDTO>
     suspend fun verifyOwnerPassword(password: String): Result<Unit>
     suspend fun changeOwnerPassword(password: String): Result<Unit>
-    suspend fun updateStorePhone(phone: String): Result<Unit>
-    suspend fun getStoreProfile(): Result<StoreProfileResponseDTO>
 }
