+++ app/src/main/java/com/gmg/seatnow/domain/usecase/auth/UpdateStorePhoneUseCase.kt
@@ -1,10 +1,11 @@
 package com.gmg.seatnow.domain.usecase.auth
 
 import com.gmg.seatnow.domain.repository.AuthRepository
+import com.gmg.seatnow.domain.repository.StoreRepository
 import javax.inject.Inject
 
 class UpdateStorePhoneUseCase @Inject constructor(
-    private val repository: AuthRepository
+    private val repository: StoreRepository
 ) {
     suspend operator fun invoke(phone: String): Result<Unit> {
         return repository.updateStorePhone(phone)
